+++ braindecode/eegneuralnet.py
@@ -20,7 +20,7 @@ from skorch.utils import noop, to_numpy, train_loss_score, valid_loss_score
 
 from braindecode.datautil import infer_signal_properties
 
-from .models.util import models_dict
+from .models.util import interpolated_models_dict, models_dict
 from .training.scoring import (
     CroppedTimeSeriesEpochScoring,
     CroppedTrialEpochScoring,
@@ -35,6 +35,8 @@ def _get_model(model: str):
     if isinstance(model, str):
         if model in models_dict:
             model = models_dict[model]
+        elif model in interpolated_models_dict:
+            model = interpolated_models_dict[model]
         else:
             raise ValueError(f"Unknown model name {model!r}.")
     return model