+++ braindecode/models/__init__.py
@@ -66,6 +66,7 @@ from .util import (
     _init_models_dict,
     build_model_config,
     extract_channel_locations_from_chs_info,
+    interpolated_models_dict,
     models_mandatory_parameters,
     positions_from_chs_info,
 )
@@ -145,4 +146,5 @@ __all__ = [
     "build_model_config",
     "_init_models_dict",
     "models_mandatory_parameters",
+    "interpolated_models_dict",
 ]